+++ tests/Menu/MenuTest.php
@@ -6,6 +6,7 @@ namespace Yiisoft\Yii\Widgets\Tests\Menu;
 
 use PHPUnit\Framework\TestCase;
 use Stringable;
+use Yiisoft\Html\IdGenerator;
 use Yiisoft\Yii\Widgets\Menu;
 use Yiisoft\Yii\Widgets\Tests\Support\Assert;
 use Yiisoft\Yii\Widgets\Tests\Support\TestTrait;
@@ -13,7 +14,9 @@ use InvalidArgumentException;
 
 final class MenuTest extends TestCase
 {
-    use TestTrait;
+    use TestTrait {
+        TestTrait::setUp as setUpTestTrait;
+    }
 
     private array $items = [
         ['label' => 'item', 'link' => '/path'],
@@ -24,6 +27,19 @@ final class MenuTest extends TestCase
         ['label' => 'Link', 'link' => '#'],
         ['label' => 'Disabled', 'link' => '#', 'disabled' => true],
     ];
+
+    protected function setUp(): void
+    {
+        $this->setUpTestTrait();
+
+        IdGenerator\disableSeed();
+        IdGenerator\reset();
+    }
+
+    protected function tearDown(): void
+    {
+        IdGenerator\enableSeed();
+    }
 
     public function testAttributes(): void
     {
@@ -314,8 +330,8 @@ final class MenuTest extends TestCase
             <ul>
             <li><a aria-current="page" class="active" href="/active">Active</a></li>
             <li>
-            <a aria-expanded="false" data-bs-toggle="dropdown" role="button" href="#">Dropdown</a>
-            <ul>
+            <a aria-expanded="false" data-bs-toggle="dropdown" role="button" id="dropdown-1" href="#">Dropdown</a>
+            <ul aria-labelledby="dropdown-1">
             <li><a href="#">Action</a></li>
             <li><a href="#">Another action</a></li>
             <li><a href="#">Something else here</a></li>
@@ -358,8 +374,8 @@ final class MenuTest extends TestCase
             <ul>
             <li><a aria-current="page" class="active" href="/active">Active</a></li>
             <li data-test="value">
-            <a aria-expanded="false" data-bs-toggle="dropdown" role="button" href="#">Dropdown</a>
-            <ul>
+            <a aria-expanded="false" data-bs-toggle="dropdown" role="button" id="dropdown-1" href="#">Dropdown</a>
+            <ul aria-labelledby="dropdown-1">
             <li><a href="#">Action</a></li>
             </ul>
             </li>
@@ -391,8 +407,8 @@ final class MenuTest extends TestCase
             <ul>
             <li><a aria-current="page" class="active" href="/active">Active</a></li>
             <li class="nav-item dropdown">
-            <a aria-expanded="false" data-bs-toggle="dropdown" role="button" href="#">Dropdown</a>
-            <ul>
+            <a aria-expanded="false" data-bs-toggle="dropdown" role="button" id="dropdown-1" href="#">Dropdown</a>
+            <ul aria-labelledby="dropdown-1">
             <li><a href="#">Action</a></li>
             <li><a href="#">Another action</a></li>
             <li><a href="#">Something else here</a></li>
@@ -436,8 +452,8 @@ final class MenuTest extends TestCase
             <ul>
             <li><a aria-current="page" class="active" href="/active">Active</a></li>
             <li>
-            <a aria-expanded="false" data-bs-toggle="dropdown" role="button" class="dropdown-toggle" href="#">Dropdown</a>
-            <ul class="dropdown-menu">
+            <a aria-expanded="false" data-bs-toggle="dropdown" role="button" class="dropdown-toggle" id="dropdown-1" href="#">Dropdown</a>
+            <ul class="dropdown-menu" aria-labelledby="dropdown-1">
             <li><a class="dropdown-item" href="#">Action</a></li>
             <li><a class="dropdown-item" href="#">Another action</a></li>
             <li><a class="dropdown-item" href="#">Something else here</a></li>